Create a simple 1-module pipeline
The `polaris.pipeline` module now provides a simple, straightforward
pipeline, with optionally a correction for induced polarization as a
result of shifts between images. Various corrections are
possible. This pipeline is then accessible as a runnable module:
`python -m polaris.pipeline`.

All options are provided through a (mandatory) configuration file in
the project root directory.